hand over position bi map to position map file
